fix: go build fixes
